error handling for resetAuth

+++ lib/util/util.dart
@@ -3,6 +3,7 @@ import 'package:path_provider/path_provider.dart';
 import 'package:ss_5/views/home.dart';
 import 'package:ss_5/communications/fileIO.dart';
 import 'package:ss_5/util/globals.dart' as globals;
+import 'dart:io';
 
 ///util contains helpful methods that are needed by many Widgets
 goHome(BuildContext context) {
@@ -16,5 +17,8 @@ resetAuth() async {
   globals.token = "";
   globals.user = null;
   String dir = (await getApplicationDocumentsDirectory()).path;
-  deleteFile('token.txt', dir);
+  File f = new File ("$dir/token.txt");
+  if(await f.exists()) {
+    deleteFile('token.txt', dir);
+  }
 }
